New dev1

+++ foodpanda-microservices-warehouse/src/main/java/com/example/foodpanda_microservices_warehouse/entity/Stock.java
@@ -17,6 +17,7 @@ import lombok.NoArgsConstructor;
 public class Stock {
     @Id
     private long id;
-    private Double stock;
+    private Integer stock;
     private String dish;
+
 }
